8315968: Move java.util.Digits to jdk.internal.util and refactor to reduce duplication
