+++ src/java.base/share/classes/jdk/internal/util/DecimalDigits.java
@@ -1,0 +1,160 @@
+/*
+ * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package jdk.internal.util;
+
+import java.lang.invoke.MethodHandle;
+
+import jdk.internal.vm.annotation.Stable;
+
+/**
+ * Digits class for decimal digits.
+ *
+ * @since 21
+ */
+public final class DecimalDigits implements Digits {
+
+    /**
+     * Each element of the array represents the packaging of two ascii characters based on little endian:<p>
+     * <pre>
+     *      00 -> '0' | ('0' << 8) -> 0x3030
+     *      01 -> '1' | ('0' << 8) -> 0x3130
+     *      02 -> '2' | ('0' << 8) -> 0x3230
+     *
+     *     ...
+     *
+     *      10 -> '0' | ('1' << 8) -> 0x3031
+     *      11 -> '1' | ('1' << 8) -> 0x3131
+     *      12 -> '2' | ('1' << 8) -> 0x3231
+     *
+     *     ...
+     *
+     *      97 -> '7' | ('9' << 8) -> 0x3739
+     *      98 -> '8' | ('9' << 8) -> 0x3839
+     *      99 -> '9' | ('9' << 8) -> 0x3939
+     * </pre>
+     */
+    @Stable
+    private static final short[] DIGITS = new short[] {
+            0x3030, 0x3130, 0x3230, 0x3330, 0x3430, 0x3530, 0x3630, 0x3730, 0x3830, 0x3930,
+            0x3031, 0x3131, 0x3231, 0x3331, 0x3431, 0x3531, 0x3631, 0x3731, 0x3831, 0x3931,
+            0x3032, 0x3132, 0x3232, 0x3332, 0x3432, 0x3532, 0x3632, 0x3732, 0x3832, 0x3932,
+            0x3033, 0x3133, 0x3233, 0x3333, 0x3433, 0x3533, 0x3633, 0x3733, 0x3833, 0x3933,
+            0x3034, 0x3134, 0x3234, 0x3334, 0x3434, 0x3534, 0x3634, 0x3734, 0x3834, 0x3934,
+            0x3035, 0x3135, 0x3235, 0x3335, 0x3435, 0x3535, 0x3635, 0x3735, 0x3835, 0x3935,
+            0x3036, 0x3136, 0x3236, 0x3336, 0x3436, 0x3536, 0x3636, 0x3736, 0x3836, 0x3936,
+            0x3037, 0x3137, 0x3237, 0x3337, 0x3437, 0x3537, 0x3637, 0x3737, 0x3837, 0x3937,
+            0x3038, 0x3138, 0x3238, 0x3338, 0x3438, 0x3538, 0x3638, 0x3738, 0x3838, 0x3938,
+            0x3039, 0x3139, 0x3239, 0x3339, 0x3439, 0x3539, 0x3639, 0x3739, 0x3839, 0x3939
+    };
+
+    /**
+     * Singleton instance of DecimalDigits.
+     */
+    public static final Digits INSTANCE = new DecimalDigits();
+
+    /**
+     * Constructor.
+     */
+    private DecimalDigits() {
+    }
+
+    @Override
+    public int digits(long value, byte[] buffer, int index,
+                      MethodHandle putCharMH) throws Throwable {
+        boolean negative = value < 0;
+        if (!negative) {
+            value = -value;
+        }
+
+        long q;
+        int r;
+        while (value <= Integer.MIN_VALUE) {
+            q = value / 100;
+            r = (int)((q * 100) - value);
+            value = q;
+            int digits = DIGITS[r];
+
+            putCharMH.invokeExact(buffer, --index, digits >> 8);
+            putCharMH.invokeExact(buffer, --index, digits & 0xFF);
+        }
+
+        int iq, ivalue = (int)value;
+        while (ivalue <= -100) {
+            iq = ivalue / 100;
+            r = (iq * 100) - ivalue;
+            ivalue = iq;
+            int digits = DIGITS[r];
+            putCharMH.invokeExact(buffer, --index, digits >> 8);
+            putCharMH.invokeExact(buffer, --index, digits & 0xFF);
+        }
+
+        if (ivalue < 0) {
+            ivalue = -ivalue;
+        }
+
+        int digits = DIGITS[ivalue];
+        putCharMH.invokeExact(buffer, --index, digits >> 8);
+
+        if (9 < ivalue) {
+            putCharMH.invokeExact(buffer, --index, digits & 0xFF);
+        }
+
+        if (negative) {
+            putCharMH.invokeExact(buffer, --index, (int)'-');
+        }
+
+        return index;
+    }
+
+    @Override
+    public int size(long value) {
+        boolean negative = value < 0;
+        int sign = negative ? 1 : 0;
+
+        if (!negative) {
+            value = -value;
+        }
+
+        long precision = -10;
+        for (int i = 1; i < 19; i++) {
+            if (value > precision)
+                return i + sign;
+
+            precision = 10 * precision;
+        }
+
+        return 19 + sign;
+    }
+
+    /**
+     * For values from 0 to 99 return a short encoding a pair of ASCII-encoded digit characters in little-endian
+     * @param i value to convert
+     * @return a short encoding a pair of ASCII-encoded digit characters
+     */
+    public static short digitPair(int i) {
+        return DIGITS[i];
+    }
+}
